refactor: move checkBox components to src/Components/CheckBox
