+++ music/src/Components/CheckBox/JazzBox.js
@@ -1,20 +1,19 @@
-import React from 'react';
-import CheckBoxCnt from './CheckBox.style';
-import '../App.css';
-import { atom } from 'recoil';
-import { useRecoilState } from 'recoil';
-import { useEffect } from 'react';
+import React, { useEffect } from 'react';
+import { atom, useRecoilState } from 'recoil';
 
-export const PopValueState = atom({
-  key: 'Pop',
+import CheckBoxCnt from './CheckBox.style';
+import '../../App.css';
+
+export const JazzValueState = atom({
+  key: 'Jazz',
   default: {},
 });
 
-const PopBox = ({ id, onNext }) => {
-  const [popValues, setPopValueState] = useRecoilState(PopValueState);
+const JazzBox = ({ id, onNext }) => {
+  const [jazzValues, setJazzValueState] = useRecoilState(JazzValueState);
 
   const checkOnlyOne = (num) => {
-    setPopValueState((prevCheckedValues) => {
+    setJazzValueState((prevCheckedValues) => {
       const updatedValues = { ...prevCheckedValues, [id]: num };
       return updatedValues;
     });
@@ -25,7 +24,7 @@ const PopBox = ({ id, onNext }) => {
 
   useEffect(() => {
     const handleCheckboxChange = () => {
-      console.log(id, popValues[id]);
+      console.log(id, jazzValues[id]);
     };
 
     const checkboxElement = document.getElementById(`checkBoxCnt_${id}`);
@@ -39,12 +38,12 @@ const PopBox = ({ id, onNext }) => {
         checkboxElement.removeEventListener('change', handleCheckboxChange);
       }
     };
-  }, [popValues, id]);
+  }, [jazzValues, id]);
 
   const generateUniqueId = (name) => `${id}_${name}`;
 
   return (
-    <CheckBoxCnt id={`checkBoxCnt_${id}`} num={popValues[id]}>
+    <CheckBoxCnt id={`checkBoxCnt_${id}`} num={jazzValues[id]}>
       {[1, 2, 3, 4, 5].map((num) => (
         <React.Fragment key={num}>
           <input
@@ -52,7 +51,7 @@ const PopBox = ({ id, onNext }) => {
             id={generateUniqueId(`b${num}`)}
             name="checkWrap"
             value={num}
-            checked={popValues[id] === num}
+            checked={jazzValues[id] === num}
             onChange={() => checkOnlyOne(num)}
           />
           <label htmlFor={generateUniqueId(`b${num}`)}>{num}</label>
@@ -62,4 +61,4 @@ const PopBox = ({ id, onNext }) => {
   );
 };
 
-export default PopBox;
+export default JazzBox;
