+++ music/src/Components/CheckBox/PopBox.js
@@ -1,20 +1,19 @@
-import React from 'react';
-import CheckBoxCnt from './CheckBox.style';
-import '../App.css';
-import { atom } from 'recoil';
-import { useRecoilState } from 'recoil';
-import { useEffect } from 'react';
+import React, { useEffect } from 'react';
+import { atom, useRecoilState } from 'recoil';
 
-export const HipValueState = atom({
-  key: 'Hip',
+import CheckBoxCnt from './CheckBox.style';
+import '../../App.css';
+
+export const PopValueState = atom({
+  key: 'Pop',
   default: {},
 });
 
-const HipBox = ({ id, onNext }) => {
-  const [hipValues, setHipValueState] = useRecoilState(HipValueState);
+const PopBox = ({ id, onNext }) => {
+  const [popValues, setPopValueState] = useRecoilState(PopValueState);
 
   const checkOnlyOne = (num) => {
-    setHipValueState((prevCheckedValues) => {
+    setPopValueState((prevCheckedValues) => {
       const updatedValues = { ...prevCheckedValues, [id]: num };
       return updatedValues;
     });
@@ -25,7 +24,7 @@ const HipBox = ({ id, onNext }) => {
 
   useEffect(() => {
     const handleCheckboxChange = () => {
-      console.log(id, hipValues[id]);
+      console.log(id, popValues[id]);
     };
 
     const checkboxElement = document.getElementById(`checkBoxCnt_${id}`);
@@ -39,12 +38,12 @@ const HipBox = ({ id, onNext }) => {
         checkboxElement.removeEventListener('change', handleCheckboxChange);
       }
     };
-  }, [hipValues, id]);
+  }, [popValues, id]);
 
   const generateUniqueId = (name) => `${id}_${name}`;
 
   return (
-    <CheckBoxCnt id={`checkBoxCnt_${id}`} num={hipValues[id]}>
+    <CheckBoxCnt id={`checkBoxCnt_${id}`} num={popValues[id]}>
       {[1, 2, 3, 4, 5].map((num) => (
         <React.Fragment key={num}>
           <input
@@ -52,7 +51,7 @@ const HipBox = ({ id, onNext }) => {
             id={generateUniqueId(`b${num}`)}
             name="checkWrap"
             value={num}
-            checked={hipValues[id] === num}
+            checked={popValues[id] === num}
             onChange={() => checkOnlyOne(num)}
           />
           <label htmlFor={generateUniqueId(`b${num}`)}>{num}</label>
@@ -62,4 +61,4 @@ const HipBox = ({ id, onNext }) => {
   );
 };
 
-export default HipBox;
+export default PopBox;
